% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMA.file.1.R
\name{COMA.file.1}
\alias{COMA.file.1}
\title{COMA.file.1}
\usage{
COMA.file.1(
  prep,
  gain.data,
  geno.data,
  pedigree.data,
  optimal.weight,
  ploidy = 2,
  map = TRUE,
  min.minor.allele = 5,
  dominance = FALSE
)
}
\arguments{
\item{prep}{StageWise prep object.}

\item{gain.data}{StageWise gain object.}

\item{geno.data}{Genotype information for StageWise.}

\item{pedigree.data}{Pedigree information for StageWise.}

\item{optimal.weight}{The optimal H Matrix blending weight.}

\item{ploidy}{Ploidy of the species.}

\item{map}{Does the Genotype data have map information for StageWise.}

\item{min.minor.allele}{How many individuals for a minor allele in StageWise.}

\item{dominance}{Should dominance effects be calculated by StageWise.}
}
\value{
File type 1 required by COMA.
}
\description{
I need a function that will create input file 1 for COMA that has the marker effects and allele dosage.
}
