% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMA.file.2.R
\name{COMA.file.2}
\alias{COMA.file.2}
\title{COMA.file.2}
\usage{
COMA.file.2(pedigree.data, ploidy = 2)
}
\arguments{
\item{pedigree.data}{A data frame with columns "id", "parent1", and "parent2" for pedigree information.}

\item{ploidy}{Integer specifying the ploidy level of the species (default is 2).}
}
\value{
A kinship matrix representing File type 2 for COMA, saved as "COMA_file_2.csv".
}
\description{
Creates input file 2 for COMA, containing the pedigree kinship matrix.
Handles missing values in \code{pedigree.data} by replacing \code{NA}s with \code{0}s.
}
