% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.lookup.R
\name{coord.lookup}
\alias{coord.lookup}
\title{coord.lookup}
\usage{
coord.lookup(Field.Map)
}
\arguments{
\item{Field.Map}{CSV file with plots laid out in field order.}
}
\value{
Dataframe containing plot numbers and their respective range and row coordinates.
}
\description{
This function assists in creating a Design File by taking a CSV map of plots arranged in field order (with cell A1 representing the range:row coordinate 1,1).
It generates a dataframe containing each plot with its corresponding range and row field coordinates.
}
