% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandData.R
\name{ExpandData}
\alias{ExpandData}
\title{ExpandData}
\usage{
ExpandData(data)
}
\arguments{
\item{data}{A data frame representing a field trial layout, with potential missing plots. Must contain columns named "Row" and "Range".}
}
\value{
A data frame with additional filler plots for a complete grid of rows and ranges.
}
\description{
Expands input data to create a complete grid of rows and ranges, filling missing plots for spatial correction.
}
