% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_markers_usef_add.R
\name{format_markers_usef_add}
\alias{format_markers_usef_add}
\title{format_markers_usef_add}
\usage{
format_markers_usef_add(geno_file, coding_type = "auto")
}
\arguments{
\item{geno_file}{File path to genotype data.}

\item{coding_type}{Original genotype coding format (0,0.5,1/-1,0,1/0,1,2/0,2) with the default of "auto" for automatic detection.}
}
\value{
A matrix object with the genotype data formatted as 0,2,NA.
}
\description{
Formats genotype data coding so that SimpleMating::getUsefA() can be ran.
}
