% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.Matrix_Optimized.R
\name{H.Matrix_Optimized}
\alias{H.Matrix_Optimized}
\title{H.Matrix_Optimized}
\usage{
H.Matrix_Optimized(
  pheno.data,
  geno.data,
  pedigree.data,
  trait,
  Fixed,
  Random,
  blend.lower = 1e-05,
  blend.upper = 0.05,
  ploidy = 2,
  max.iter = 100,
  map = TRUE,
  dominance = FALSE,
  workspace = "12Gb",
  pworkspace = "12Gb",
  min.minor.allele = 5,
  fix.eff.marker = NULL,
  covariates = NULL,
  mask = NULL,
  method = NULL,
  numCores = 2,
  weight.vector.length = 2,
  non.add = "none",
  max_global_size = 1e+10
)
}
\arguments{
\item{pheno.data}{Phenotype data for StageWise.}

\item{geno.data}{Genotype data for StageWise.}

\item{pedigree.data}{Pedigree data for StageWise.}

\item{trait}{Character of trait(s) to pass StageWise.}

\item{Fixed}{Fixed effects to pass StageWise.}

\item{Random}{Random effects to pass StageWise.}

\item{blend.lower}{The lower limit of the A matrix contribution for blending.}

\item{blend.upper}{The upper limit of the A matrix contribution for blending.}

\item{ploidy}{Ploidy of the species.}

\item{max.iter}{The maxium number of iterations for ASReml.}

\item{map}{Whether the genotype data includes map coordinate infromation in the first three columns.}

\item{dominance}{Should dominance effects be calculated with StageWise.}

\item{workspace}{Memory allocation arguement for ASReml.}

\item{pworkspace}{Memory allocation arguement for ASReml.}

\item{min.minor.allele}{How many individuals must have the minor allele for StageWise.}

\item{fix.eff.marker}{Marker effect arguement for StageWise.}

\item{covariates}{Covariate arguement for StageWise.}

\item{mask}{Should StageWise mask some of the data.}

\item{method}{Arguement for StageWise.}

\item{numCores}{How many cores should be used for parrele processing.}

\item{weight.vector.length}{How many different blending weights should be tested.}

\item{non.add}{How should non additive genetic effects be handled.}

\item{max_global_size}{Memory allocation arguement.}
}
\value{
StageWise prep object that has the optimal blending of the H Matrix.
}
\description{
This function tests the proportion of blending to perform in generating the H Matrix so as to minimize the AIC. (If the model does not converge it is more likely an issue of the range of weight values being tested for blending than that of not having the max.iter high enough. If the model does not converge in less than 50 iterations there is generally a bigger issue at hand.)
}
