% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.master.R
\name{harvest.master}
\alias{harvest.master}
\title{harvest.master}
\usage{
harvest.master(
  Location,
  Design.File,
  Combine.File,
  Fixed = c("id_1"),
  Random,
  Moisture = 15.5,
  Bushel = 56,
  Plot.Length = 22.5,
  Row.Spacing = 30,
  Plot.Rows = 2,
  nest.terms = NULL,
  nest.name = NULL
)
}
\arguments{
\item{Location}{Location name for the data.}

\item{Design.File}{Filename for the Design.File.}

\item{Combine.File}{Filename for the Combine.File.}

\item{Fixed}{Character for Fixed Terms, "id_1" is the default.}

\item{Random}{Character for Random Terms.}

\item{Moisture}{Numeric for the standard moisture content.}

\item{Bushel}{Numeric for the standard weight of a bushel.}

\item{Plot.Length}{Numeric for the length of the plot in feet.}

\item{Row.Spacing}{Numeric for the spacing of rows in inches.}

\item{Plot.Rows}{Numeric for the number of rows in the plot.}

\item{nest.terms}{List of character vectors to be nested.}

\item{nest.name}{Character vector or new nesting term name(s).}
}
\value{
Dataframe that is properly formated and has had the yield calculated.
}
\description{
This function will take the raw output file directly off of a H2 Harvest Master system with the Mirus Software and apply proper formatting so that the output is ready for StageWise. The user must provide a Design File that has columns "plot", "range", "row" that ties the Harvest Master output to that of the respective experimental design features for said plot such as Block. User must specify the location (environment).
}
