% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insilico.R
\name{insilico}
\alias{insilico}
\title{insilico}
\usage{
insilico(
  geno_file_female,
  geno_file_male,
  output_file = "insilico_hybrids.csv",
  coding_type = "auto"
)
}
\arguments{
\item{geno_file_female}{File path to the female parent inbreds' genotype data.}

\item{geno_file_male}{File path to the male parent inbreds' genotype data.}

\item{output_file}{File path to save the hybrid genotypes (CSV).}

\item{coding_type}{Original genotype coding format (0,0.5,1/-1,0,1/0,1,2/0,2) with default "auto".}
}
\value{
NULL (writes output to file instead of returning a data frame).
}
\description{
Generates in silico hybrid genotypes from inbred parents and writes them to a file.
File is written incrementally to disk as opposed to being stored in RAM. This reduces the speed of the operation but is required.
}
\details{
The total memory needed is: Memory = (number of females)×(number of males)×(number of markers)×8bytes. This shows why writing to
disk is required as RAM requirements scales quickly (2570\emph{1525}2490*8 ~ 72.7 Gb).
}
