% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insilico_transposed.R
\name{insilico_transposed}
\alias{insilico_transposed}
\title{insilico_transposed}
\usage{
insilico_transposed(
  geno_file_female,
  geno_file_male,
  output_file = "insilico_hybrids_transposed.csv",
  coding_type = "auto"
)
}
\arguments{
\item{geno_file_female}{File path to the female parent inbreds' genotype data.}

\item{geno_file_male}{File path to the male parent inbreds' genotype data.}

\item{output_file}{File path to save the hybrid genotypes (CSV), defaults to "insilico_hybrids_transposed.csv".}

\item{coding_type}{Original genotype coding format (0,0.5,1/-1,0,1/0,1,2/0,2) with default "auto".}
}
\value{
NULL (writes output to file instead of returning a data frame).
}
\description{
Generates in silico hybrid genotypes from inbred parents and writes them to a file in a transposed format,
with hybrids as rows and markers as columns. The reason the in silico hybrids are being regenerated as opposed
to just using the existing in silico hybrids file is that the prediction stage is set up to use the in silico
information in a transposed manner. Making the hybrids the rows was a choice as it is assumed that the user will
have many more hybrids than markers. This makes the predictions much more efficient (long vs wide format).
For large datasets, RAM limitations prevent simply taking the transpose of the existing in silico file.
The file is written incrementally to disk to handle large datasets efficiently, avoiding high memory usage.
}
\details{
The total memory requirement for the full dataset would be approximately:
Memory = (number of females) × (number of males) × (number of markers) × 8 bytes.
For example, 2570 females × 1525 males × 2490 markers × 8 bytes ~ 72.7 GB, making incremental writing essential.
}
