% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.terms.R
\name{model.terms}
\alias{model.terms}
\title{model.terms}
\usage{
model.terms(Fixed, Random)
}
\arguments{
\item{Fixed}{Character vector of the fixed effects for the model (e.g., c("env", "block")).}

\item{Random}{Character vector of the random effects for the model (e.g., c("genotype")).}
}
\value{
A data frame with columns for the term names, their fixed/random designation, and factor status,
formatted for StageWise compatibility.
}
\description{
Prepares the model terms required for StageWise functions, ensuring that the formatting aligns with the specific syntax
expectations of StageWise. This function allows users to specify fixed and random effects in a straightforward way,
and it handles the necessary formatting automatically.
}
