% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.terms_update.R
\name{model.terms_update}
\alias{model.terms_update}
\title{model.terms_update}
\usage{
model.terms_update(df, Design.File, Fixed, Random)
}
\arguments{
\item{Fixed}{Character of the Fixed Effects for your model.}

\item{Random}{Character of the Random Effects for your model.}
}
\value{
Dataframe that helps handle StageWise model syntax.
}
\description{
The input file with the phenotypic information from the Harvest Master does not contain all the experimental design features and desired covariates for the model. The Model Term columns must be updated with the corresponding values, they cannot just remain NA, this is done with the Design File. The Design File must contain columns for; env, plot, range and row.
}
