% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.R
\name{nesting}
\alias{nesting}
\title{nesting}
\usage{
nesting(df, nest.terms, nest.name)
}
\arguments{
\item{df}{Data frame containing the original data.}

\item{nest.terms}{List of character vectors, where each vector contains column names to be nested together.}

\item{nest.name}{Character vector of new names for the nested terms.}
}
\value{
A data frame with nested terms created and original columns removed.
}
\description{
This function creates composite "nested" terms by merging specified columns in \code{nest.terms} and assigning them a new
name specified by \code{nest.name}. This allows StageWise to handle nested terms in the format required for modeling.
For example, if Subblock 5 in Block 2 is nested, it can be represented as "Subblock_5_Block_2".
}
