% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nursery_constraints.R
\name{nursery_constraints}
\alias{nursery_constraints}
\title{nursery_constraints}
\usage{
nursery_constraints(df, nursery.rows)
}
\arguments{
\item{df}{Data frame from OMA with the proposed matings. Must include a \code{value} column indicating percent contribution of each cross.}

\item{nursery.rows}{Integer specifying the total nursery rows available for allocation.}
}
\value{
A data frame indicating each cross and the number of nursery rows allocated to each.
}
\description{
This function allocates nursery rows to each biparental cross based on their percent contribution (\code{value}) to the next cycle.
Allocation stops once the specified number of \code{nursery.rows} has been reached.
This function is used for either the DH induction stage or the ear-to-row generational advancement stage, where generation of bi-parental crosses is not limiting.
}
