% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma_optimized.R
\name{oma_optimized}
\alias{oma_optimized}
\title{oma_optimized}
\usage{
oma_optimized(
  dF,
  parents,
  matings,
  ploidy,
  K,
  tol = 1e-06,
  dF.adapt = NULL,
  solver = "ECOS",
  base = "current",
  sparse_threshold = 1000
)
}
\arguments{
\item{dF}{Targeted inbreeding rate.}

\item{parents}{Data frame containing the parents.}

\item{matings}{Data frame containing specified matings.}

\item{ploidy}{Numeric value for ploidy.}

\item{K}{Kinship matrix.}

\item{tol}{Tolerance threshold.}

\item{dF.adapt}{Allows for inbreeding rate adjustment.}

\item{solver}{Solver method for CVXR.}

\item{base}{Population type.}

\item{sparse_threshold}{Number of crosses to begin using sparse matrices, user option allows trade off for speed vs RAM efficiency.}
}
\value{
A list containing response data.frame, optimal contributions data.frame, and optimal allocations data.frame.
}
\description{
Optimizes COMA::oma for use with large datasets. The key difference is the use of sparse matrices as opposed to dense matrices.
This allows for more efficient RAM allocation enabling the use of larger datasets.
Additionally, R has a 32-bit integer limit of 2^31-1 so if ((# parents) * (# matings)) >= 2^31-1,
then matings is filtered to maximize the number of matings evaluated while remaining below the 32-bit limit (issue resides in CVXR solver).
}
