% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma_reduced.R
\name{oma_reduced}
\alias{oma_reduced}
\title{oma_reduced}
\usage{
oma_reduced(
  dF = c(0.01, 0.01),
  geno.file,
  kinship.file,
  ploidy = 2,
  selection.intensity = 0.1,
  solver = "ECOS",
  dF.adapt = list(step = 0.005, max = 1),
  base = "RM"
)
}
\arguments{
\item{geno.file}{Genotype information for COMA.}

\item{kinship.file}{Kinship matrix for COMA.}

\item{ploidy}{Ploidy of species.}

\item{selection.intensity}{The top percent of parents to consider using for large sets of germplasm.}

\item{solver}{The solver COMA should utilize.}

\item{dF.adapt}{If the specified dF does not find a solution how should COMA proceed.}

\item{base}{Are the parents in question inbreds and COMA should use random mating.}

\item{df}{Inbreeding rate constraints.}
}
\value{
OMA output.
}
\description{
When the number of parents results in too many pairwise mating combinations to evaluate (due to RAM), this function will first perform OCS and identify the top parental candidates. Then using the OCS output OMA will be conducted in order to find the best matings.
}
