% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(data, trait)
}
\arguments{
\item{data}{Raw input file to be cleaned.}

\item{trait}{Character of trait(s) to be cleaned in the input file.}
}
\value{
Dataframe that is properly formated and has had the yield calculated.
}
\description{
Interactive detection of outliers based on histograms. Writes a new dataframe with erroneous data containing plots set to NA. Multiple traits can be evaluated at once using trait = c().
}
