% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planCross.R
\name{planCross}
\alias{planCross}
\title{planCross}
\usage{
planCross(TargetPop, MateDesign = "half", TargetPop2 = NULL, Indiv2keep = NULL)
}
\arguments{
\item{TargetPop}{Parents for crossing.}

\item{MateDesign}{Mating design choice.}

\item{TargetPop2}{Parents to cross with TargetPop.}

\item{Indiv2keep}{Filter for parents to use.}
}
\value{
A data.frame object with both parents for a cross.
}
\description{
Optimizes SimpleMating::planCross to handle large crossing designs more efficiently. The primary difference is the vectorization of building crosses and not iteratively building the crossing schedule.
}
