% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coordinates.R
\name{plot.coordinates}
\alias{plot.coordinates}
\title{plot.coordinates}
\usage{
\method{plot}{coordinates}(Field.Map)
}
\arguments{
\item{Field.Map}{CSV file of plots laid out as the field is.}
}
\value{
Dataframe of the plots and their range and row coordinates.
}
\description{
This function can help to part in create a Design File as it takes a CSV Map of Plots laid out in field order (cell A1 is range:row coordinate 1,1) and generates a dataframe with the plot and its' corresponding range and row field coordinates.
}
