% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.Pedigree.R
\name{PRISM.Pedigree}
\alias{PRISM.Pedigree}
\title{PRISM.Pedigree}
\usage{
PRISM.Pedigree(data)
}
\arguments{
\item{data}{A CSV file containing the PRISM-exported pedigree history.}
}
\value{
A three-column pedigree dataframe with the columns: \code{id} (individual), \code{parent1} (female parent), and \code{parent2} (male parent).
}
\description{
PRISM is a database system commonly used in the commercial sector for pedigree tracking. This function processes an export from PRISM to create a standardized three-column pedigree dataframe
(individual, female parent, male parent) suitable for generating a kinship matrix with COMA.file.2.
The PRISM export must contain the columns "Pedigree", "PedId", "Female PedId", and "Male PedId".
}
