% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_optimized.R
\name{read_data_optimized}
\alias{read_data_optimized}
\title{read_data_optimized}
\usage{
read_data_optimized(
  geno.file,
  kinship.file,
  ploidy,
  sex = NULL,
  matings = "none",
  standardize = FALSE,
  n.core = 4,
  partition = FALSE,
  chunk_size = 1e+05,
  result_file = "matings_results.csv"
)
}
\arguments{
\item{geno.file}{Character. Path to the genotype file for COMA.}

\item{kinship.file}{Character. Path to the kinship matrix for COMA.}

\item{ploidy}{Numeric. Ploidy level of the species.}

\item{sex}{Dataframe or NULL. Contains 'id' and 'female' columns for COMA (optional).}

\item{matings}{Character or Dataframe. Specifies mating pairs; can be "none" or "all" for default options.}

\item{standardize}{Logical. If TRUE, standardizes the genetic merit.}

\item{n.core}{Integer. Number of cores for parallel processing.}

\item{partition}{Logical. Argument passed to COMA.}

\item{chunk_size}{Integer. Number of mating pairs to process per core in each chunk.}

\item{result_file}{Character. Filename for output.}
}
\value{
List containing the kinship matrix \code{K}, parental data \code{parents}, and mating data \code{matings}.
}
\description{
This function optimizes the \code{read_data()} functionality from COMA for large datasets by processing mating merit calculations in chunks,
reducing memory usage by saving intermediary results to disk. The output is compatible with COMA's kinship matrix and genetic merit calculations.
}
