% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1_plots.R
\name{Stage1_plots}
\alias{Stage1_plots}
\title{Stage1_plots}
\usage{
Stage1_plots(filename, trait, Fixed, Random)
}
\arguments{
\item{filename}{Character string specifying the path to the phenotype data file for StageWise.}

\item{trait}{Character string specifying a single trait to be analyzed in StageWise.}

\item{Fixed}{Formula or list of fixed effects to be passed to StageWise.}

\item{Random}{Formula or list of random effects to be passed to StageWise.}
}
\value{
A list containing four ggplot objects:
\itemize{
\item \code{boxplot}: Boxplot of model residuals.
\item \code{qqplot}: QQ plot of model residuals.
\item \code{heritability}: Boxplot of broad-sense heritability (H^2) by environment.
\item \code{spatial}: Spatial field map of residuals and trends.
}
}
\description{
Generate diagnostic plots for the Stage 1 model fitted to a single trait using the SPATS solver.
The function creates a boxplot of residuals, a QQ plot, a heritability plot, and a spatial field map.
}
