% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMA.file.1.R
\name{COMA.file.1}
\alias{COMA.file.1}
\title{COMA.file.1}
\usage{
COMA.file.1(
  prep,
  gain.data,
  geno.data,
  pedigree.data,
  optimal.weight,
  ploidy = 2,
  map = TRUE,
  min.minor.allele = 5,
  dominance = FALSE
)
}
\arguments{
\item{prep}{A data frame or list containing preprocessed data for StageWise analysis.}

\item{gain.data}{A data frame with gain data including index coefficients and trait information.}

\item{geno.data}{Genotype information used by StageWise.}

\item{pedigree.data}{Pedigree information for individuals in StageWise.}

\item{optimal.weight}{A numeric value for the optimal H Matrix blending weight.}

\item{ploidy}{Integer specifying the ploidy level of the species (default is 2).}

\item{map}{Logical; if TRUE, genotype data includes map information for StageWise.}

\item{min.minor.allele}{Minimum count of individuals required for a minor allele.}

\item{dominance}{Logical; if TRUE, calculates dominance effects in addition to additive effects.}
}
\value{
A data frame representing File type 1 required by COMA, containing marker effects.
}
\description{
Creates input file 1 for COMA, containing marker effects and allele dosage information.
}
