% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMA.file.2.R
\name{COMA.file.2}
\alias{COMA.file.2}
\title{COMA.file.2}
\usage{
COMA.file.2(pedigree.data, ploidy = 2)
}
\arguments{
\item{pedigree.data}{Pedigree information for StageWise.}

\item{ploidy}{Ploidy of the species.}
}
\value{
File type 2 required by COMA.
}
\description{
I need a function that will create input file 2 for COMA that has the pedigree kinship matrix. The only modification to Amatrix() is that the NA handling and converting to 0 is handled by this function. The input pedigree.data is made outside the function as naming conventions (text separators) are not consistent, requiring the user to make a dataframe with three columns c("id", "parent1", "parent2").
}
