% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandData.R
\name{ExpandData}
\alias{ExpandData}
\title{ExpandData}
\usage{
ExpandData(data)
}
\arguments{
\item{data}{Input data that has plots missing.}
}
\value{
Input data with filler plots added where range/row were incomplete.
}
\description{
Makes a full grid so that spatial correction is possible. This is necessary as the ranges and rows of a field trial are not always perfectly filled with plots. In order for spatial correction to work with ASReml R, these ranges and rows must be complete so "filler" plots are created to satisfy this.
}
