% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.Matrix_Optimized.R
\name{H.Matrix_Optimized}
\alias{H.Matrix_Optimized}
\title{H.Matrix_Optimized}
\usage{
H.Matrix_Optimized(
  pheno.data,
  geno.data,
  pedigree.data,
  trait,
  Fixed,
  Random,
  blend.lower = 1e-05,
  blend.upper = 0.2,
  ploidy = 2,
  max.iter = 100,
  map = TRUE,
  dominance = FALSE,
  workspace = "12Gb",
  pworkspace = "12Gb",
  min.minor.allele = 5,
  fix.eff.marker = NULL,
  covariates = NULL,
  mask = NULL,
  method = NULL,
  non.add = "none",
  tol = 1e-04
)
}
\arguments{
\item{pheno.data}{Data frame containing phenotype data for StageWise.}

\item{geno.data}{Data frame containing genotype data for StageWise.}

\item{pedigree.data}{Data frame containing pedigree data for StageWise.}

\item{trait}{Character vector of trait(s) to be passed to StageWise.}

\item{Fixed}{Formula or list of fixed effects to be passed to StageWise.}

\item{Random}{Formula or list of random effects to be passed to StageWise.}

\item{blend.lower}{Lower limit of the A matrix contribution for blending (default is 1e-5).}

\item{blend.upper}{Upper limit of the A matrix contribution for blending (default is 0.20).}

\item{ploidy}{Integer specifying the ploidy level of the species (default is 2).}

\item{max.iter}{Maximum number of iterations for ASReml (default is 100).}

\item{map}{Logical; TRUE if genotype data includes map coordinates in the first three columns (default is TRUE).}

\item{dominance}{Logical; TRUE if dominance effects should be calculated with StageWise (default is FALSE).}

\item{workspace}{Memory allocation for ASReml (default is "12Gb").}

\item{pworkspace}{Memory allocation for ASReml's parallel workspace (default is "12Gb").}

\item{min.minor.allele}{Minimum number of individuals with the minor allele for StageWise (default is 5).}

\item{fix.eff.marker}{Marker effect argument for StageWise (default is NULL).}

\item{covariates}{Covariate argument for StageWise (default is NULL).}

\item{mask}{Logical; if TRUE, StageWise masks some of the data (default is NULL).}

\item{method}{Method argument for StageWise (default is NULL).}

\item{non.add}{Specifies handling of non-additive genetic effects (default is "none").}

\item{tol}{Tolerance level for optimization convergence (default is 1e-4).}
}
\value{
A StageWise prep object containing the optimal H Matrix with minimized AIC and BLUPs prepared.
}
\description{
Optimizes the blending weight for generating the H Matrix to minimize the AIC, by testing various proportions
of blending. This function is designed to help select an optimal blending weight for a mixed model in StageWise
based on AIC. Convergence issues are generally related to the blending weight range, not iteration limits.
}
