% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.master.R
\name{harvest.master}
\alias{harvest.master}
\title{harvest.master}
\usage{
harvest.master(
  Location,
  Design.File,
  Combine.File,
  Fixed = c("id_1"),
  Random,
  Moisture = 15.5,
  Bushel = 56,
  Plot.Length = 22.5,
  Row.Spacing = 30,
  Plot.Rows = 2,
  nest.terms = NULL,
  nest.name = NULL
)
}
\arguments{
\item{Location}{Character; the name of the location (environment) for the data.}

\item{Design.File}{Filename (or path) for the design file containing experimental design details.}

\item{Combine.File}{Filename (or path) for the raw Harvest Master output file.}

\item{Fixed}{Character vector for fixed terms; "id_1" is the default.}

\item{Random}{Character vector for random terms.}

\item{Moisture}{Numeric; the standard moisture content to adjust weights (default is 15.5).}

\item{Bushel}{Numeric; standard weight of a bushel (default is 56).}

\item{Plot.Length}{Numeric; the length of each plot in feet (default is 22.5).}

\item{Row.Spacing}{Numeric; spacing of rows in inches (default is 30).}

\item{Plot.Rows}{Numeric; the number of rows in each plot (default is 2).}

\item{nest.terms}{List of character vectors specifying terms to be nested.}

\item{nest.name}{Character vector specifying new name(s) for nesting term(s).}
}
\value{
A data frame formatted for StageWise, with calculated yield values.
}
\description{
Processes raw output from an H2 Harvest Master system with Mirus Software, applying formatting and calculations to
prepare the data for use in StageWise. Requires a design file with columns "plot", "range", and "row" that links
the Harvest Master data to experimental design features such as blocks. The location (environment) of the data
collection must be specified.
}
