% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.terms.R
\name{model.terms}
\alias{model.terms}
\title{model.terms}
\usage{
model.terms(Fixed, Random)
}
\arguments{
\item{Fixed}{Character of the Fixed Effects for your model.}

\item{Random}{Character of the Random Effects for your model.}
}
\value{
Dataframe that helps handle StageWise model syntax.
}
\description{
StageWise requires the model terms be passed to the various functions in specific formatting. This includes the input data files. This function allows the user to enter the model in a more straight forward manner and have the formatting handled for you.
}
