% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.terms_update.R
\name{model.terms_update}
\alias{model.terms_update}
\title{model.terms_update}
\usage{
model.terms_update(df, Design.File, Fixed, Random)
}
\arguments{
\item{df}{Data frame containing the initial phenotypic information from the Harvest Master system.}

\item{Design.File}{Filename (or path) of the design file containing the design features and covariates.}

\item{Fixed}{Character vector specifying the fixed effects for the model.}

\item{Random}{Character vector specifying the random effects for the model.}
}
\value{
A data frame with updated model terms, including design features and covariates for StageWise compatibility.
}
\description{
Updates the model term columns in the input data frame by joining it with a design file that contains experimental design
features and covariates. This is necessary as the raw phenotypic data file from Harvest Master may lack these
experimental design features (e.g., range, row, etc.) which are required for model building.
}
\details{
The design file must include columns: "env", "plot", "range", and "row".
}
