% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.R
\name{nesting}
\alias{nesting}
\title{nesting}
\usage{
nesting(df, nest.terms, nest.name)
}
\arguments{
\item{df}{Origin data.}

\item{nest.terms}{List of character vectors to be nested.}

\item{nest.name}{Character vector or new nesting term name(s).}
}
\value{
Dataframe that helps handle StageWise model syntax.
}
\description{
In order to handle nesting terms in StageWise you must create a new "composite" term. (Ex. Subblock 5 in Block 2 becomes one term "Subblock_5_Block_2") Therefore I need to handle performing the merging and removal of the old terms. This needs to be handled within the harvest.master function. nest.terms is a list of character vectors and nest.name is just a character vector.
}
