% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nursery_constraints.R
\name{nursery_constraints}
\alias{nursery_constraints}
\title{nursery_constraints}
\usage{
nursery_constraints(df, nursery.rows)
}
\arguments{
\item{df}{Dataframe from OMA with the proposed matings.}
}
\value{
Crosses and how many nursery rows should be allocated to each.
}
\description{
oma and conversely oma_reduced returns the percent contribution for each cross to the next cycle. In a breeding program you are constrained by available resources in regards to the number of available rows. This returns the allocation of rows to each biparental cross at either the DH induction stage or the ear to row generational advancement stage. It is assumed that the generation of the bi-parental crosses is not a limiting factor.
}
