% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma_reduced.R
\name{oma_reduced}
\alias{oma_reduced}
\title{oma_reduced}
\usage{
oma_reduced(
  dF = c(0.01, 0.01),
  geno.file,
  kinship.file,
  ploidy = 2,
  selection.intensity = 0.1,
  solver = "ECOS",
  dF.adapt = list(step = 0.005, max = 1),
  base = "RM"
)
}
\arguments{
\item{dF}{Numeric vector of length 2 indicating inbreeding rate constraints for OMA. Defaults to c(0.01, 0.01).}

\item{geno.file}{File path for genotype data to be used in OMA.}

\item{kinship.file}{File path for kinship matrix to be used in OMA.}

\item{ploidy}{Integer indicating ploidy level of the species. Defaults to 2.}

\item{selection.intensity}{Numeric value representing the proportion of top parents to select based on OCS. Defaults to 0.10 (top 10\%).}

\item{solver}{Character string specifying the solver COMA should utilize, e.g., "ECOS".}

\item{dF.adapt}{List specifying adaptive parameters for \code{dF} if the initial value does not yield a solution (e.g., list(step = 0.005, max = 1)).}

\item{base}{Character indicating if the base population consists of random matings ("RM") or inbred lines.}
}
\value{
A data frame containing the OMA output with optimal mating combinations.
}
\description{
This function performs an optimized mating allocation (OMA) with reduced computational demands by selecting a top subset of parental candidates via optimal contribution selection (OCS) before running OMA. This is useful when a high number of parent combinations would otherwise exceed memory constraints.
}
