% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(data, trait)
}
\arguments{
\item{data}{Dataframe containing raw input data to be cleaned.}

\item{trait}{Character vector of trait(s) to be cleaned in the input file.}
}
\value{
Dataframe with outliers set to NA for the specified traits.
}
\description{
Detects outliers interactively by displaying histograms of specified traits, allowing the user to set upper and lower bounds iteratively.
The function updates the data by removing outliers and sets the values of flagged outlier data to NA.
Multiple traits can be evaluated simultaneously by passing them as a character vector in \code{trait}.
}
