% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.Pedigree.R
\name{PRISM.Pedigree}
\alias{PRISM.Pedigree}
\title{PRISM.Pedigree}
\usage{
PRISM.Pedigree(data)
}
\arguments{
\item{data}{CSV of PRISM exported pedigree history.}
}
\value{
Three column pedigree dataframe properly formated.
}
\description{
PRISM is the standard database system in the commercial sector (or a variant based off of it). PRISM is capable of exporting Pedigree History from a defined History Start. This function will allow you to create a complete three column pedigree dataframe from the PRISM export to use with COMA.file.2 for generating a pedigree kinship matrix. The PRISM export must contain the following columns; "Pedigree", "PedId", "Female PedId", and "Male PedId".
}
